refactor(ui): assets themes
