+++ packages/twenty-ui/assets/themes/dark-noise.jpg
