+++ packages/twenty-ui/assets/themes/light-noise.png
