+++ packages/twenty-ui/src/theme/constants/BackgroundDark.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @nx/workspace-no-hardcoded-colors */
-import DarkNoise from '../assets/dark-noise.jpg';
+import DarkNoise from '@/assets/themes/dark-noise.jpg';
 
 import { COLOR } from './Colors';
 import { GRAY_SCALE } from './GrayScale';
